requirements.txt
bidict==0.21.2
click==7.1.2
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
openpyxl==3.0.7
pandas==1.1.5
Pillow==8.1.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-engineio==4.0.1
python-socketio==5.1.0
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.23
Werkzeug==1.0.1
